Update webpack build files
